fix msg actuator
