+++ dapp-chain/oracle/src/main/java/org/tron/service/eventactuator/MsgActuator.java
@@ -1,0 +1,8 @@
+package org.tron.service.eventactuator;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j(topic = "actuator")
+public abstract class MsgActuator extends Actuator {
+
+}
