+++ dapp-chain/oracle/src/main/java/org/tron/service/eventactuator/msgactuator/mainchain/DepositTRC10MsgActuator.java
@@ -1,4 +1,4 @@
-package org.tron.service.eventmsgactuator.mainchain;
+package org.tron.service.eventactuator.msgactuator.mainchain;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
@@ -18,7 +18,7 @@ import org.tron.protos.Sidechain.EventMsg;
 import org.tron.protos.Sidechain.EventMsg.EventType;
 import org.tron.protos.Sidechain.EventMsg.TaskEnum;
 import org.tron.service.capsule.TransactionExtensionCapsule;
-import org.tron.service.eventmsgactuator.MsgActuator;
+import org.tron.service.eventactuator.MsgActuator;
 
 @Slf4j(topic = "mainChainTask")
 public class DepositTRC10MsgActuator extends MsgActuator {
@@ -36,8 +36,7 @@ public class DepositTRC10MsgActuator extends MsgActuator {
     ByteString valueBS = ByteString.copyFrom(ByteArray.fromString(value));
     ByteString nonceBS = ByteString.copyFrom(ByteArray.fromString(nonce));
     this.event = DepositTRC10Event.newBuilder().setFrom(fromBS).setValue(valueBS)
-        .setTokenId(tokenIdBS)
-        .setNonce(nonceBS).build();
+        .setTokenId(tokenIdBS).setNonce(nonceBS).build();
   }
 
   public DepositTRC10MsgActuator(EventMsg eventMsg) throws InvalidProtocolBufferException {
@@ -67,10 +66,9 @@ public class DepositTRC10MsgActuator extends MsgActuator {
             assetIssue.getName().toStringUtf8(), assetIssue.getName().toStringUtf8(),
             assetIssue.getPrecision());
       }
-      Transaction tx = SideChainGatewayApi
-          .mintToken10Transaction(fromStr, tokenIdStr, valueStr,
-              assetIssue.getName().toStringUtf8(),
-              assetIssue.getName().toStringUtf8(), assetIssue.getPrecision(), nonceStr);
+      Transaction tx = SideChainGatewayApi.mintToken10Transaction(fromStr, tokenIdStr, valueStr,
+          assetIssue.getName().toStringUtf8(), assetIssue.getName().toStringUtf8(),
+          assetIssue.getPrecision(), nonceStr);
       this.transactionExtensionCapsule = new TransactionExtensionCapsule(NONCE_TAG + nonceStr, tx,
           0);
       return CreateRet.SUCCESS;
