+++ dapp-chain/oracle/src/main/java/org/tron/service/eventactuator/msgactuator/mainchain/DepositTRC721MsgActuator.java
@@ -1,4 +1,4 @@
-package org.tron.service.eventmsgactuator.mainchain;
+package org.tron.service.eventactuator.msgactuator.mainchain;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
@@ -6,40 +6,42 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import java.util.Objects;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.tron.client.MainChainGatewayApi;
 import org.tron.client.SideChainGatewayApi;
 import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.WalletUtil;
 import org.tron.core.net.message.EventNetMessage;
 import org.tron.protos.Protocol.Transaction;
+import org.tron.protos.Sidechain.DepositTRC721Event;
 import org.tron.protos.Sidechain.EventMsg;
 import org.tron.protos.Sidechain.EventMsg.EventType;
 import org.tron.protos.Sidechain.EventMsg.TaskEnum;
-import org.tron.protos.Sidechain.MappingTRC20Event;
 import org.tron.service.capsule.TransactionExtensionCapsule;
-import org.tron.service.eventmsgactuator.MsgActuator;
+import org.tron.service.eventactuator.MsgActuator;
 
 @Slf4j(topic = "mainChainTask")
-public class MappingTRC20MsgActuator extends MsgActuator {
+public class DepositTRC721MsgActuator extends MsgActuator {
 
-  private static final String NONCE_TAG = "mapping_";
+  private static final String NONCE_TAG = "deposit_";
 
-  private MappingTRC20Event event;
+  private DepositTRC721Event event;
   @Getter
-  private EventType type = EventType.MAPPING_TRC20;
+  private EventType type = EventType.DEPOSIT_TRC721_EVENT;
   @Getter
   private TaskEnum taskEnum = TaskEnum.SIDE_CHAIN;
 
-  public MappingTRC20MsgActuator(String contractAddress, String nonce) {
+  public DepositTRC721MsgActuator(String from, String contractAddress, String uid,
+      String nonce) {
+    ByteString fromBS = ByteString.copyFrom(WalletUtil.decodeFromBase58Check(from));
     ByteString contractAddressBS = ByteString
         .copyFrom(WalletUtil.decodeFromBase58Check(contractAddress));
+    ByteString uidBS = ByteString.copyFrom(ByteArray.fromString(uid));
     ByteString nonceBS = ByteString.copyFrom(ByteArray.fromString(nonce));
-    this.event = MappingTRC20Event.newBuilder().setContractAddress(contractAddressBS)
-        .setNonce(nonceBS).build();
+    this.event = DepositTRC721Event.newBuilder().setFrom(fromBS).setUId(uidBS)
+        .setContractAddress(contractAddressBS).setNonce(nonceBS).build();
   }
 
-  public MappingTRC20MsgActuator(EventMsg eventMsg) throws InvalidProtocolBufferException {
-    this.event = eventMsg.getParameter().unpack(MappingTRC20Event.class);
+  public DepositTRC721MsgActuator(EventMsg eventMsg) throws InvalidProtocolBufferException {
+    this.event = eventMsg.getParameter().unpack(DepositTRC721Event.class);
   }
 
   @Override
@@ -48,19 +50,17 @@ public class MappingTRC20MsgActuator extends MsgActuator {
       return CreateRet.SUCCESS;
     }
     try {
+      String fromStr = WalletUtil.encode58Check(event.getFrom().toByteArray());
       String contractAddressStr = WalletUtil
           .encode58Check(event.getContractAddress().toByteArray());
+      String uIdStr = event.getUId().toStringUtf8();
       String nonceStr = event.getNonce().toStringUtf8();
 
-      long trcDecimals = MainChainGatewayApi.getTRCDecimals(contractAddressStr);
-      String trcName = MainChainGatewayApi.getTRCName(contractAddressStr);
-      String trcSymbol = MainChainGatewayApi.getTRCSymbol(contractAddressStr);
       logger.info(
-          "MappingTRC20Event, contractAddress: {}, trcName: {}, trcSymbol: {}, trcDecimals: {}, nonce: {}.",
-          contractAddressStr, trcName, trcSymbol, trcDecimals, nonceStr);
-
+          "DepositTRC721Actuator, from: {}, tokenId: {}, contractAddress: {}, nonce: {}",
+          fromStr, uIdStr, contractAddressStr, nonceStr);
       Transaction tx = SideChainGatewayApi
-          .multiSignForMappingTRC20(contractAddressStr, trcName, trcSymbol, trcDecimals, nonceStr);
+          .mintToken721Transaction(fromStr, contractAddressStr, uIdStr, nonceStr);
       this.transactionExtensionCapsule = new TransactionExtensionCapsule(NONCE_TAG + nonceStr, tx,
           0);
       return CreateRet.SUCCESS;
@@ -89,17 +89,13 @@ public class MappingTRC20MsgActuator extends MsgActuator {
   @Override
   public EventNetMessage generateSignedEventMsg() {
 
+    String fromStr = WalletUtil.encode58Check(event.getFrom().toByteArray());
     String contractAddressStr = WalletUtil
         .encode58Check(event.getContractAddress().toByteArray());
+    String valueStr = event.getUId().toStringUtf8();
     String nonceStr = event.getNonce().toStringUtf8();
-
-    long trcDecimals = MainChainGatewayApi.getTRCDecimals(contractAddressStr);
-    String trcName = MainChainGatewayApi.getTRCName(contractAddressStr);
-    String trcSymbol = MainChainGatewayApi.getTRCSymbol(contractAddressStr);
     return SideChainGatewayApi
-        .getMappingTRC20SignMsg(contractAddressStr, trcName, trcSymbol, trcDecimals, nonceStr,
-            getMessage());
-
+        .getTRCSignMsg(fromStr, contractAddressStr, valueStr, nonceStr, getMessage());
   }
 
 }
