+++ dapp-chain/oracle/src/main/java/org/tron/service/eventactuator/msgactuator/sidechain/WithdrawTRXMsgActuator.java
@@ -1,4 +1,4 @@
-package org.tron.service.eventmsgactuator.mainchain;
+package org.tron.service.eventactuator.msgactuator.sidechain;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
@@ -11,37 +11,33 @@ import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.WalletUtil;
 import org.tron.core.net.message.EventNetMessage;
 import org.tron.protos.Protocol.Transaction;
-import org.tron.protos.Sidechain.DepositTRC20Event;
 import org.tron.protos.Sidechain.EventMsg;
 import org.tron.protos.Sidechain.EventMsg.EventType;
 import org.tron.protos.Sidechain.EventMsg.TaskEnum;
+import org.tron.protos.Sidechain.WithdrawTRXEvent;
 import org.tron.service.capsule.TransactionExtensionCapsule;
-import org.tron.service.eventmsgactuator.MsgActuator;
+import org.tron.service.eventactuator.MsgActuator;
 
-@Slf4j(topic = "mainChainTask")
-public class DepositTRC20MsgActuator extends MsgActuator {
+@Slf4j(topic = "sideChainTask")
+public class WithdrawTRXMsgActuator extends MsgActuator {
 
-  private static final String NONCE_TAG = "deposit_";
-
-  private DepositTRC20Event event;
+  private static final String PREFIX = "withdraw_1_";
+  private WithdrawTRXEvent event;
   @Getter
-  private EventType type = EventType.DEPOSIT_TRC20_EVENT;
+  private EventType type = EventType.WITHDRAW_TRX_EVENT;
   @Getter
   private TaskEnum taskEnum = TaskEnum.SIDE_CHAIN;
 
-  public DepositTRC20MsgActuator(String from, String contractAddress, String value,
-      String nonce) {
+  public WithdrawTRXMsgActuator(String from, String value, String nonce) {
     ByteString fromBS = ByteString.copyFrom(WalletUtil.decodeFromBase58Check(from));
     ByteString valueBS = ByteString.copyFrom(ByteArray.fromString(value));
-    ByteString contractAddressBS = ByteString
-        .copyFrom(WalletUtil.decodeFromBase58Check(contractAddress));
     ByteString nonceBS = ByteString.copyFrom(ByteArray.fromString(nonce));
-    this.event = DepositTRC20Event.newBuilder().setFrom(fromBS).setValue(valueBS)
-        .setContractAddress(contractAddressBS).setNonce(nonceBS).build();
+    this.event = WithdrawTRXEvent.newBuilder().setFrom(fromBS).setValue(valueBS).setNonce(nonceBS)
+        .build();
   }
 
-  public DepositTRC20MsgActuator(EventMsg eventMsg) throws InvalidProtocolBufferException {
-    this.event = eventMsg.getParameter().unpack(DepositTRC20Event.class);
+  public WithdrawTRXMsgActuator(EventMsg eventMsg) throws InvalidProtocolBufferException {
+    this.event = eventMsg.getParameter().unpack(WithdrawTRXEvent.class);
   }
 
   @Override
@@ -51,18 +47,15 @@ public class DepositTRC20MsgActuator extends MsgActuator {
     }
     try {
       String fromStr = WalletUtil.encode58Check(event.getFrom().toByteArray());
-      String contractAddressStr = WalletUtil
-          .encode58Check(event.getContractAddress().toByteArray());
       String valueStr = event.getValue().toStringUtf8();
       String nonceStr = event.getNonce().toStringUtf8();
 
-      logger.info("DepositTRC20Actuator, from: {}, value: {}, contractAddress: {}, nonce: {}",
-          fromStr, valueStr, contractAddressStr, nonceStr);
+      logger
+          .info("WithdrawTRXActuator, from: {}, value: {}, nonce: {}", fromStr, valueStr, nonceStr);
 
       Transaction tx = SideChainGatewayApi
-          .mintToken20Transaction(fromStr, contractAddressStr, valueStr, nonceStr);
-      this.transactionExtensionCapsule = new TransactionExtensionCapsule(NONCE_TAG + nonceStr, tx,
-          0);
+          .withdrawTRXTransaction(fromStr, valueStr, nonceStr);
+      this.transactionExtensionCapsule = new TransactionExtensionCapsule(PREFIX + nonceStr, tx, 0);
       return CreateRet.SUCCESS;
     } catch (Exception e) {
       logger.error("when create transaction extension capsule", e);
@@ -78,7 +71,7 @@ public class DepositTRC20MsgActuator extends MsgActuator {
 
   @Override
   public byte[] getNonceKey() {
-    return ByteArray.fromString(NONCE_TAG + event.getNonce().toStringUtf8());
+    return ByteArray.fromString(PREFIX + event.getNonce().toStringUtf8());
   }
 
   @Override
@@ -90,12 +83,8 @@ public class DepositTRC20MsgActuator extends MsgActuator {
   public EventNetMessage generateSignedEventMsg() {
 
     String fromStr = WalletUtil.encode58Check(event.getFrom().toByteArray());
-    String contractAddressStr = WalletUtil
-        .encode58Check(event.getContractAddress().toByteArray());
     String valueStr = event.getValue().toStringUtf8();
     String nonceStr = event.getNonce().toStringUtf8();
-    return SideChainGatewayApi
-        .getTRCSignMsg(fromStr, contractAddressStr, valueStr, nonceStr, getMessage());
+    return SideChainGatewayApi.getTRXSignMsg(fromStr, valueStr, nonceStr, getMessage());
   }
-
 }
