--- dapp-chain/oracle/src/main/java/org/tron/service/eventmsgactuator/SignListParam.java
@@ -1,17 +1,0 @@
-package org.tron.service.eventmsgactuator;
-
-import java.util.List;
-import lombok.Getter;
-
-public class SignListParam {
-
-  @Getter
-  private List<String> oracleSigns;
-  @Getter
-  private List<String> oracleAddresses;
-
-  public SignListParam(List<String> oracleSigns, List<String> oracleAddresses) {
-    this.oracleSigns = oracleSigns;
-    this.oracleAddresses = oracleAddresses;
-  }
-}
